Update the way example output is generated
Make use of 'examples' in the yaml file to provide the updated examples.
This also allows the ability to provide multiple examples, i.e. for
search results. #15

Rather than being generated from the component parts, the updated
examples are provided as blocks of json.

The updated examples are also provided as 'ttl' and 'xml'. These were
generated from the json examples.

Updated examples are provided for:
- /artefacts
- /artefacts/{artefactID}
- /artefacts/{artefactID}/record
- /artefacts/{artefactID}/resources
- /artefacts/{artefactID}/resources/{resourceID}
- /artefacts/{artefactID}/resources/classes
- /artefacts/{artefactID}/resources/concepts
- /artefacts/{artefactID}/resources/properties
- /artefacts/{artefactID}/resources/individuals
- /artefacts/{artefactID}/resources/schemes
- /artefacts/{artefactID}/resources/collections
- /records
- /records/{artefactID}
- /search
- /search/content
- /search/metadata

For the updated examples the prefixes have been removed #24.

Additionally '@id' was added to:
- skos:Collection
- skos:Concept
- skos:ConceptScheme